# docs/swagger.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: ISecL Workload Service
servers:
  - url: http://workload-service.com/wls
paths:
...
  /images/{id}:
    get:
      tags:
        - Images
      summary: Retrieve an Image 
      description: Query an Image by its UUID and see it's associated flavorId
      parameters:
        - $ref : '#/components/parameters/imageIdParam'
      responses:
        '200':
          description: Successfully retrieved Image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          description: Image not found
    delete:
      tags:
        - Images
      summary: Delete an Image
      description: Delete an Image by its UUID
      parameters:
        - $ref: '#/components/parameters/imageIdParam'
      responses:
        '204':
          description: Successfully deleted Image
        '404':
          description: Image not found
...
components:
...
  parameters:
...
    imageIdParam:
      name: id
      in: path
      description: Image ID
      required: true
      schema:
        type: string
        format: uuid
    # reportIdParam:
    #   name: id
    #   in: path
    #   description: Report ID
    #   required: true
    #   schema: 
    #     type: integer
    #     format: int64
    # userIdParam:
    #   name: id
    #   in: path
    #   description: User id
    #   required: true
    #   schema:
    #     type: integer
    #     format: int64
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Image:
      type: object
      required:
        - id
        - flavor_id
      properties:
        id:
          type: string
          format: uuid
        flavor_id:
          type: string
          format: uuid
...